export validation functions
